# backend/openapi_debug_2.json
{"openapi":"3.1.0","info":{"title":"CollectSecure API","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/compliance/check-call-window/{zip_code}":{"get":{"tags":["Compliance"],"summary":"Api Check Call Window","operationId":"api_check_call_window_compliance_check_call_window__zip_code__get","parameters":[{"name":"zip_code","in":"path","required":true,"schema":{"type":"string","title":"Zip Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/upload":{"post":{"tags":["Ingest"],"summary":"Upload Portfolio","description":"Accepts a CSV file upload and processes it in the background.","operationId":"upload_portfolio_api_v1_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_portfolio_api_v1_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/verify-epay":{"get":{"tags":["Operations"],"summary":"Verify Epay","description":"Diagnostic endpoint to verify USA ePay connectivity.","operationId":"verify_epay_api_v1_verify_epay_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/portfolios":{"get":{"tags":["Operations"],"summary":"Get Portfolios","operationId":"get_portfolios_api_v1_portfolios_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/work-queue":{"get":{"tags":["Operations"],"summary":"Get Work Queue","description":"Fetches the next available debts for the agent.\nLogic: Query 'debts' where status='New' LIMIT 1","operationId":"get_work_queue_api_v1_work_queue_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/DebtResponse"},"type":"array","title":"Response Get Work Queue Api V1 Work Queue Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/search":{"get":{"tags":["Operations"],"summary":"Search Debts","description":"Search for debts by name or client reference.\nsearch_type: 'name' or 'client_ref'\nquery: search term","operationId":"search_debts_api_v1_search_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"search_type","in":"query","required":true,"schema":{"type":"string","title":"Search Type"}},{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DebtResponse"},"title":"Response Search Debts Api V1 Search Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/interactions":{"post":{"tags":["Operations"],"summary":"Log Interaction","description":"Logs a call/email. Checks compliance before logging.","operationId":"log_interaction_api_v1_interactions_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InteractionCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/email/send":{"post":{"tags":["Operations"],"summary":"Send Template Email","description":"Send a template-based email for a specific debt.","operationId":"send_template_email_api_v1_email_send_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailTemplateSend"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/debts/{debt_id}/email":{"put":{"tags":["Operations"],"summary":"Update Debtor Email","description":"Update debtor email for a specific debt.","operationId":"update_debtor_email_api_v1_debts__debt_id__email_put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"debt_id","in":"path","required":true,"schema":{"type":"integer","title":"Debt Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DebtorEmailUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/payments":{"post":{"tags":["Operations"],"summary":"Process Payment","description":"Processes a payment and splits the ledger.","operationId":"process_payment_api_v1_payments_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/payments/{debt_id}":{"get":{"tags":["Operations"],"summary":"Get Debt Payments","description":"Fetches all payments recorded for a specific debt.","operationId":"get_debt_payments_api_v1_payments__debt_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"debt_id","in":"path","required":true,"schema":{"type":"integer","title":"Debt Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PaymentResponse"},"title":"Response Get Debt Payments Api V1 Payments  Debt Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/payment-plans/preview":{"get":{"tags":["Operations"],"summary":"Get Payment Plan Preview","description":"Returns a preview of a payment schedule without saving it.","operationId":"get_payment_plan_preview_api_v1_payment_plans_preview_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"total_amount","in":"query","required":true,"schema":{"anyOf":[{"type":"number"},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"}],"title":"Total Amount"}},{"name":"down_payment","in":"query","required":true,"schema":{"anyOf":[{"type":"number"},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"}],"title":"Down Payment"}},{"name":"installments","in":"query","required":true,"schema":{"type":"integer","title":"Installments"}},{"name":"frequency","in":"query","required":true,"schema":{"type":"string","title":"Frequency"}},{"name":"start_date","in":"query","required":true,"schema":{"type":"string","format":"date-time","title":"Start Date"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":true},"title":"Response Get Payment Plan Preview Api V1 Payment Plans Preview Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/payment-plans":{"post":{"tags":["Operations"],"summary":"Create Payment Plan","description":"Creates a new payment plan and generates its schedule.","operationId":"create_payment_plan_api_v1_payment_plans_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentPlanCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentPlanResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/payment-plans/{debt_id}":{"get":{"tags":["Operations"],"summary":"Get Debt Plans","operationId":"get_debt_plans_api_v1_payment_plans__debt_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"debt_id","in":"path","required":true,"schema":{"type":"integer","title":"Debt Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PaymentPlanResponse"},"title":"Response Get Debt Plans Api V1 Payment Plans  Debt Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/payment-plans/{plan_id}/schedule":{"get":{"tags":["Operations"],"summary":"Get Plan Schedule","operationId":"get_plan_schedule_api_v1_payment_plans__plan_id__schedule_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"plan_id","in":"path","required":true,"schema":{"type":"integer","title":"Plan Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ScheduledPaymentResponse"},"title":"Response Get Plan Schedule Api V1 Payment Plans  Plan Id  Schedule Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/debug/epay-account":{"get":{"tags":["Operations"],"summary":"Debug Epay Account","operationId":"debug_epay_account_api_v1_debug_epay_account_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/debug/run-next-installment":{"post":{"tags":["Operations"],"summary":"Debug Run Next Installment","operationId":"debug_run_next_installment_api_v1_debug_run_next_installment_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"plan_id","in":"query","required":true,"schema":{"type":"integer","title":"Plan Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/payments/scheduled/{payment_id}/execute":{"post":{"tags":["Operations"],"summary":"Execute Scheduled Payment","description":"Manually executes a scheduled payment early.","operationId":"execute_scheduled_payment_api_v1_payments_scheduled__payment_id__execute_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"payment_id","in":"path","required":true,"schema":{"type":"integer","title":"Payment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/payments":{"get":{"tags":["Operations"],"summary":"Get Admin Payments","description":"Admin view for payment management.","operationId":"get_admin_payments_api_v1_admin_payments_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"days","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Days"}},{"name":"start_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Start Date"}},{"name":"end_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"End Date"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/payments/one-off":{"post":{"tags":["Operations"],"summary":"Run One Off Payment","description":"Runs a manual payment using the most recent card token for this debt.","operationId":"run_one_off_payment_api_v1_payments_one_off_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"debt_id","in":"query","required":true,"schema":{"type":"integer","title":"Debt Id"}},{"name":"amount","in":"query","required":true,"schema":{"anyOf":[{"type":"number"},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"}],"title":"Amount"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reports/daily-money":{"get":{"tags":["Operations"],"summary":"Get Daily Money Report","operationId":"get_daily_money_report_api_v1_reports_daily_money_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"date","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reports/liquidation":{"get":{"tags":["Operations"],"summary":"Get Liquidation Report","operationId":"get_liquidation_report_api_v1_reports_liquidation_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"portfolio_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Portfolio Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/webhooks/sendgrid":{"post":{"tags":["Webhooks","webhooks"],"summary":"Sendgrid Webhook","description":"Handle SendGrid webhook events for email tracking.\n\nEvents include: delivered, opened, clicked, bounced, dropped, spam_report, unsubscribe\n\nSendGrid webhook documentation:\nhttps://docs.sendgrid.com/for-developers/tracking-events/event","operationId":"sendgrid_webhook_api_webhooks_sendgrid_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/webhooks/sendgrid/test":{"get":{"tags":["Webhooks","webhooks"],"summary":"Test Webhook","description":"Test endpoint to verify webhook is accessible.","operationId":"test_webhook_api_webhooks_sendgrid_test_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/campaigns/templates":{"get":{"tags":["Campaigns","Campaigns"],"summary":"List Templates","operationId":"list_templates_api_v1_campaigns_templates_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["Campaigns","Campaigns"],"summary":"Register Template","operationId":"register_template_api_v1_campaigns_templates_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TemplateCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/campaigns/preview":{"post":{"tags":["Campaigns","Campaigns"],"summary":"Preview Audience","operationId":"preview_audience_api_v1_campaigns_preview_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampaignFilter"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampaignPreview"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/campaigns/launch":{"post":{"tags":["Campaigns","Campaigns"],"summary":"Launch Campaign","description":"Creates a campaign and triggers background sending.","operationId":"launch_campaign_api_v1_campaigns_launch_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampaignCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/campaigns":{"get":{"tags":["Campaigns","Campaigns"],"summary":"List Campaigns","operationId":"list_campaigns_api_v1_campaigns_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}}},"components":{"schemas":{"Body_upload_portfolio_api_v1_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_portfolio_api_v1_upload_post"},"CampaignCreate":{"properties":{"name":{"type":"string","title":"Name"},"subject":{"type":"string","title":"Subject"},"template_id":{"type":"string","title":"Template Id"},"filters":{"$ref":"#/components/schemas/CampaignFilter"}},"type":"object","required":["name","subject","template_id","filters"],"title":"CampaignCreate"},"CampaignFilter":{"properties":{"min_balance":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Min Balance"},"max_balance":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Max Balance"},"portfolio_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Portfolio Id"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"},"last_email_status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Email Status"},"last_email_before":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Email Before"},"last_email_after":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Email After"},"last_email_older_than_days":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Last Email Older Than Days"},"include_unemailed":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Include Unemailed","default":false}},"type":"object","title":"CampaignFilter"},"CampaignPreview":{"properties":{"recipient_count":{"type":"integer","title":"Recipient Count"}},"type":"object","required":["recipient_count"],"title":"CampaignPreview"},"DebtResponse":{"properties":{"client_reference_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Reference Number"},"original_account_number":{"type":"string","title":"Original Account Number"},"original_creditor":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Original Creditor"},"date_opened":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date Opened"},"charge_off_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Charge Off Date"},"principal_balance":{"anyOf":[{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Principal Balance"},"fees_costs":{"anyOf":[{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Fees Costs"},"amount_due":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Amount Due"},"last_payment_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Payment Date"},"last_payment_amount":{"anyOf":[{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Last Payment Amount"},"total_paid_amount":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Total Paid Amount","default":"0.00"},"id":{"type":"integer","title":"Id"},"status":{"type":"string","title":"Status"},"debtor":{"$ref":"#/components/schemas/DebtorResponse"}},"type":"object","required":["original_account_number","amount_due","id","status","debtor"],"title":"DebtResponse"},"DebtorEmailUpdate":{"properties":{"email":{"type":"string","title":"Email"}},"type":"object","required":["email"],"title":"DebtorEmailUpdate"},"DebtorResponse":{"properties":{"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"dob":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dob"},"address_1":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address 1"},"address_2":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address 2"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"},"zip_code":{"type":"string","title":"Zip Code"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"mobile_consent":{"type":"boolean","title":"Mobile Consent","default":false},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"id":{"type":"string","title":"Id"},"ssn_hash":{"type":"string","title":"Ssn Hash"},"do_not_contact":{"type":"boolean","title":"Do Not Contact"}},"type":"object","required":["first_name","last_name","zip_code","id","ssn_hash","do_not_contact"],"title":"DebtorResponse"},"EmailTemplateSend":{"properties":{"debt_id":{"type":"integer","title":"Debt Id"},"template_id":{"type":"string","title":"Template Id"}},"type":"object","required":["debt_id","template_id"],"title":"EmailTemplateSend"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InteractionCreate":{"properties":{"debt_id":{"type":"integer","title":"Debt Id"},"action_type":{"type":"string","title":"Action Type"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","required":["debt_id","action_type"],"title":"InteractionCreate"},"PaymentCreate":{"properties":{"debt_id":{"type":"integer","title":"Debt Id"},"amount_paid":{"anyOf":[{"type":"number"},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"}],"title":"Amount Paid"}},"type":"object","required":["debt_id","amount_paid"],"title":"PaymentCreate"},"PaymentPlanCreate":{"properties":{"debt_id":{"type":"integer","title":"Debt Id"},"total_settlement_amount":{"anyOf":[{"type":"number"},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"}],"title":"Total Settlement Amount"},"is_settlement":{"type":"boolean","title":"Is Settlement","default":true},"down_payment_amount":{"anyOf":[{"type":"number"},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"}],"title":"Down Payment Amount","default":"0.00"},"installment_count":{"type":"integer","title":"Installment Count"},"frequency":{"type":"string","title":"Frequency"},"start_date":{"type":"string","format":"date-time","title":"Start Date"},"payment_key":{"type":"string","title":"Payment Key"},"cardholder_name":{"type":"string","title":"Cardholder Name"},"billing_address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Billing Address"},"billing_city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Billing City"},"billing_state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Billing State"},"billing_zip":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Billing Zip"}},"type":"object","required":["debt_id","total_settlement_amount","installment_count","frequency","start_date","payment_key","cardholder_name"],"title":"PaymentPlanCreate"},"PaymentPlanResponse":{"properties":{"id":{"type":"integer","title":"Id"},"debt_id":{"type":"integer","title":"Debt Id"},"total_settlement_amount":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Total Settlement Amount"},"is_settlement":{"type":"boolean","title":"Is Settlement"},"down_payment_amount":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Down Payment Amount"},"installment_count":{"type":"integer","title":"Installment Count"},"frequency":{"type":"string","title":"Frequency"},"start_date":{"type":"string","format":"date-time","title":"Start Date"},"status":{"type":"string","title":"Status"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"down_payment_status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Down Payment Status"},"dp_reference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dp Reference"}},"type":"object","required":["id","debt_id","total_settlement_amount","is_settlement","down_payment_amount","installment_count","frequency","start_date","status","created_at"],"title":"PaymentPlanResponse"},"PaymentResponse":{"properties":{"id":{"type":"integer","title":"Id"},"amount_paid":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Amount Paid"},"agency_portion":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Agency Portion"},"client_portion":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Client Portion"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"transaction_reference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Transaction Reference"},"scheduled_payment_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Scheduled Payment Id"},"payment_method":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Payment Method"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"},"result_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Result Code"},"result":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Result"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"}},"type":"object","required":["id","amount_paid","agency_portion","client_portion","timestamp"],"title":"PaymentResponse"},"ScheduledPaymentResponse":{"properties":{"id":{"type":"integer","title":"Id"},"amount":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Amount"},"due_date":{"type":"string","format":"date-time","title":"Due Date"},"status":{"type":"string","title":"Status"},"actual_payment_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Actual Payment Id"},"processed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Processed At"},"transaction_reference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Transaction Reference"},"payment_method":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Payment Method"},"attempt_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Attempt Count"},"next_attempt_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Next Attempt At"},"last_result_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Result Code"},"last_result":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Result"},"last_decline_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Decline Reason"},"last_error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Error"}},"type":"object","required":["id","amount","due_date","status"],"title":"ScheduledPaymentResponse"},"TemplateCreate":{"properties":{"name":{"type":"string","title":"Name"},"template_id":{"type":"string","title":"Template Id"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["name","template_id"],"title":"TemplateCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}